+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG for husky
 
 * dev-develop
+    * ENHANCEMENT #700 added datagrid get-records and delete-records event
     * BUGFIX      #708 Only show draft and publish icons in column navigation when page is no ghost
 
 * 0.24.0 (2016-08-08)